+++ packages/framework/bundle/domain/bundle.js
@@ -56,24 +56,6 @@ function() {
 	"protocol" : ["Oskari.bundle.Bundle"],
 	"source" : {
 		"scripts" : [{
-            "type" : "text/javascript",
-            "src" : "../../../../sources/framework/domain/AbstractLayer.js"
-        }, {
-            "type" : "text/javascript",
-            "src" : "../../../../sources/framework/domain/wmslayer.js"
-        }, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/domain/vectorlayer.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/domain/map.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/domain/style.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/domain/tool.js"
-		},{
 			"type" : "text/javascript",
 			"src" : "../../../../sources/framework/domain/user.js"
 		}],